pyproject.toml
[project]
name = "flaxchat"
version = "0.1.0"
description = "A minimal end-to-end LLM training harness for TPUs, built on JAX/Flax NNX"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Taha Bouhsine" },
]
dependencies = [
    "jax>=0.6.0",
    "flax>=0.11.0",
    "optax>=0.2.4",
    "orbax-checkpoint>=0.11.0",
    "datasets>=4.0.0",
    "tokenizers>=0.22.0",
    "tiktoken>=0.11.0",
    "rustbpe>=0.1.0",
    "fastapi>=0.117.1",
    "uvicorn>=0.36.0",
    "wandb>=0.21.3",
    "pyarrow>=18.0.0",
    "pyyaml>=6.0",
    "filelock>=3.0",
]

[project.optional-dependencies]
tpu = ["jax[tpu]>=0.6.0", "tpuz>=0.1.0"]
kaggle = ["kgz>=0.1.0"]
dev = [
    "pytest>=8.0.0",
    "matplotlib>=3.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
